third_party/WebKit/LayoutTests/http/tests/w3c/webperf/submission/Intel/user-timing/test_user_timing_method_exist-expected.txt

Description

This test validates that all of the user timing methods exist in window.performance (but does not validate that their functionality).


PASS window.performance is defined 
PASS performance.mark is supported! 
PASS performance.clearMarks is supported! 
PASS performance.measure is supported! 
PASS performance.clearMeasures is supported! 
PASS performance.getEntries is supported! 
PASS performance.getEntriesByType is supported! 
PASS performance.getEntriesByName is supported! 
